Added Login Page
I have added Login. Now need to make the login page react to the option chosen prior visiting the login page. Also need to change the nav bar view in the login page and make all the elements align perfectly. 
I have also added the nav bar and the footer to the login page as well. 

I anyone feels like there needs to be a change. Please download this repository and modify accordingly.